+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsParamRepository.kt
@@ -10,7 +10,20 @@ class AnalyticsParamRepository(
     private val uuidHelper: UUIDHelper = UUIDHelper()
 ) {
 
+    /**
+     * Link type used for navigating back to the merchant app. See [LinkType].
+     */
     var linkType: LinkType? = null
+
+    /**
+     * App switch enabled by the merchant request
+     */
+    var didEnablePayPalAppSwitch: Boolean? = null
+
+    /**
+     * App switch attempted based on the server side response
+     */
+    var didPayPalServerAttemptAppSwitch: Boolean? = null
 
     private lateinit var _sessionId: String
 
@@ -26,10 +39,17 @@ class AnalyticsParamRepository(
         }
 
     /**
-     * Clears the session ID value from the repository
+     * Resets the [sessionId] and clears all other repository values.
+     *
+     * Note that this function is called in different spots of the SDK lifecycle for different payment modules. Some
+     * modules call reset during launch of the SDK. The PayPal module calls reset at the end of the payment flow to
+     * persist the [sessionId] value set from the Shopper Insights module.
      */
-    fun resetSessionId() {
+    fun reset() {
         _sessionId = uuidHelper.formattedUUID
+        linkType = null
+        didEnablePayPalAppSwitch = null
+        didPayPalServerAttemptAppSwitch = null
     }
 
     companion object {