+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalCheckoutRequestUnitTest.java
@@ -283,8 +283,8 @@ public class PayPalCheckoutRequestUnitTest {
     @Test
     public void createRequestBody_sets_contactInformation_when_not_null() throws JSONException {
         PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00", true);
-
         request.setContactInformation(new PayPalContactInformation("some@email.com", new PayPalPhoneNumber("1", "1234567890")));
+
         String requestBody = request.createRequestBody(
                 mock(Configuration.class),
                 mock(Authorization.class),
@@ -311,5 +311,36 @@ public class PayPalCheckoutRequestUnitTest {
 
         assertFalse(requestBody.contains("\"recipient_email\":\"some@email.com\""));
         assertFalse(requestBody.contains("\"international_phone\":{\"country_code\":\"1\",\"national_number\":\"1234567890\"}"));
+    }
+
+    @Test
+    public void createRequestBody_sets_contactPreference_when_not_null() throws JSONException {
+        PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00", true);
+        request.setContactPreference(PayPalContactPreference.UPDATE_CONTACT_INFORMATION);
+
+        String requestBody = request.createRequestBody(
+            mock(Configuration.class),
+            mock(Authorization.class),
+            "success_url",
+            "cancel_url",
+            null
+        );
+
+        assertTrue(requestBody.contains("\"contact_preference\":\"UPDATE_CONTACT_INFO\""));
+    }
+
+    @Test
+    public void createRequestBody_does_not_set_contactPreference_when_null() throws JSONException {
+        PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00", true);
+
+        String requestBody = request.createRequestBody(
+            mock(Configuration.class),
+            mock(Authorization.class),
+            "success_url",
+            "cancel_url",
+            null
+        );
+
+        assertFalse(requestBody.contains("contact_preference"));
     }
 }