+++ TestUtils/src/main/java/com/braintreepayments/api/testutils/MockBraintreeClientBuilder.java
@@ -19,6 +19,11 @@ import com.braintreepayments.api.sharedutils.HttpResponseCallback;
 
 import org.mockito.stubbing.Answer;
 
+/**
+ * This outputs a Mockito mock of BraintreeClient.
+ * For writing tests in kotlin, use {@link com.braintreepayments.api.testutils.MockkBraintreeClientBuilder} that outputs a mockk mock.
+ */
+@Deprecated
 public class MockBraintreeClientBuilder {
 
     private String sendGETSuccess;