+++ Venmo/src/main/java/com/braintreepayments/api/venmo/VenmoClient.kt
@@ -375,25 +375,35 @@ class VenmoClient internal constructor(
 
     private fun callbackPaymentAuthFailure(
         callback: VenmoPaymentAuthRequestCallback,
-        request: VenmoPaymentAuthRequest
+        request: VenmoPaymentAuthRequest.Failure
     ) {
-        braintreeClient.sendAnalyticsEvent(VenmoAnalytics.TOKENIZE_FAILED, analyticsParams)
+        braintreeClient.sendAnalyticsEvent(
+            VenmoAnalytics.TOKENIZE_FAILED,
+            analyticsParams.copy(errorDescription = request.error.message)
+        )
         callback.onVenmoPaymentAuthRequest(request)
+        analyticsParamRepository.reset()
     }
 
     private fun callbackSuccess(callback: VenmoTokenizeCallback, venmoResult: VenmoResult) {
         braintreeClient.sendAnalyticsEvent(VenmoAnalytics.TOKENIZE_SUCCEEDED, analyticsParams)
         callback.onVenmoResult(venmoResult)
+        analyticsParamRepository.reset()
     }
 
     private fun callbackTokenizeCancel(callback: VenmoTokenizeCallback) {
         braintreeClient.sendAnalyticsEvent(VenmoAnalytics.APP_SWITCH_CANCELED, analyticsParams)
         callback.onVenmoResult(VenmoResult.Cancel)
+        analyticsParamRepository.reset()
     }
 
-    private fun callbackTokenizeFailure(callback: VenmoTokenizeCallback, venmoResult: VenmoResult) {
-        braintreeClient.sendAnalyticsEvent(VenmoAnalytics.TOKENIZE_FAILED, analyticsParams)
+    private fun callbackTokenizeFailure(callback: VenmoTokenizeCallback, venmoResult: VenmoResult.Failure) {
+        braintreeClient.sendAnalyticsEvent(
+            VenmoAnalytics.TOKENIZE_FAILED,
+            analyticsParams.copy(errorDescription = venmoResult.error.message)
+        )
         callback.onVenmoResult(venmoResult)
+        analyticsParamRepository.reset()
     }
 
     private val analyticsParams: AnalyticsEventParams