+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Braintree Android SDK Release Notes
 
-## unreleased
+## 5.9.0 (2025-04-09)
 
-* PayPal
-  * Make `PayPalCheckoutRequest.userAuthenticationEmail` optional for App Switch flows
+* All Modules
   * Add `error_desc` tag to the analytics events we are sending to FPTI.
+* PayPal
+  * Send `merchant_enabled_app_switch` and `attempted_app_switch` in `event_params` for App Switch events to PayPal's analytics service (FPTI)
+  * Add `contactPreference` property to `PayPalCheckoutRequest` - This property is required to enable the contact information section in the payment flow.
+  * Make `PayPalCheckoutRequest.userAuthenticationEmail` optional for App Switch flows
 * ThreeDSecure
   * Bump Cardinal version to `2.2.7-7` (fixes #1236).